requirements.txt
google-genai
requests
beautifulsoup4
markdown
python-dotenv
